Please enter the commit message for your changes. Lines starting

+++ models/dual_encoder.py
@@ -55,7 +55,7 @@ def dual_encoder_model(
     # Run the utterance and context through the RNN
     rnn_outputs, rnn_states = tf.nn.dynamic_rnn(
         cell,
-        tf.concat(0, [context_embedded, context_token_len_avg, context_nums, utterance_embedded, utterance_token_len_avg, utterance_nums]),
+        tf.concat(0, [context_embedded, tf.float32(context_token_len_avg), tf.float32(context_nums), utterance_embedded, tf.float32(utterance_token_len_avg), tf.float32(utterance_nums)]),
         sequence_length=tf.concat(0, [context_len+2, utterance_len+2]),
         dtype=tf.float32)
     encoding_context, encoding_utterance = tf.split(0, 2, rnn_states.h)
